workers/pyproject.toml
[project]
name = "raijin-workers"
version = "0.1.0"
description = "Raijin workers — OCR, validation, notifications"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4",
    "redis>=5.2",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "azure-ai-documentintelligence>=1.0.0",
    "boto3>=1.35",
    "python-dateutil>=2.9",
    "structlog>=24.4",
    "sentry-sdk[celery]>=2.19",
    "cryptography>=43.0",
    "msal>=1.31",
    "requests>=2.32",
    "google-auth>=2.36",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.150",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
